rag_service/requirements.txt
redis
python-dotenv
requests
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
docling
